non member attr/funct
member, non member function / attr.

member attr/func:  (instance var and function) only access in the instance its in.
use this-> from the inside of the class and [instance].XXX from the outside

non member func/attr: (class vars and functions) static functions that are the same in the whole class.
directly use the class name and the :: (scope operator) just like namespaces.

Initialize nonmember attribute  when code starts running to assign it a value, otherwise it will be an undefined state.
for that we'll add the class definition in the .cpp (Sample::_nbInst = 0;) < only way to initialize a non member attribute.